Convert to WebExtension
Perform the step to a WebExtension addon. This falls into two parts: A
background script and a content script for the message displaying tab.

The background script reacts on the command key, injecting the content
script into the current tab that displays a message. It furthermore
waits for messages sent from the tab by the content script.

The content script requests via a message from the background script the
currently selected email message. This is needed as the content script
has no access to the CopyPatch API experiment. It then extract the patch
from the message, displays warnings as they are detected, and then sends
the patch text via a message to the background script for injection into
the clipboard. Again, this is needed as the content script has no access
to the related API.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>